Moved environment-variables to userland.
Now that we have the new UserAccess-concept, it is more
complicated than before. Additionally, if we do the
absolutize-path in userland, we don't need any env-vars in
kernel anymore. Of course, we might do additional work by
first absolutizing the path in userland and then copying
it again in kernel, but because the UserAccess stuff costs
a bit, too, we are actually not worse.

+++ source/include/sys/arch/i586/task/uenv.h
@@ -28,7 +28,7 @@ class UEnv : public UEnvBase {
 
 	static void startSignalHandler(Thread *t,IntrptStackFrame *stack,int sig,
 	                               Signals::handler_func handler);
-	static uint32_t *addArgs(Thread *t,uint32_t *esp,uintptr_t tentryPoint,bool newThread);
+	static ulong *addArgs(Thread *t,ulong *esp,uintptr_t tentryPoint,bool newThread);
 };
 
 inline void UEnvBase::handleSignal(Thread *t,IntrptStackFrame *stack) {
